Genlog: Enable options at start of command line
All the first command line arguments starting with '-' (hyphen) are
passed on to the various test scripts. Some scripts may get more options
than others in the future, but all of them accept the -t/--todo and
-a/--all options. So:

  Genlog -t *.t  # Run TODO tests
  Genlog -a *.t  # Run all tests

Several options can be specified at the start of the command line, but
in this case it doesn't make much sense to mix -a and -t.

f24a20d8-3693-11e5-ba45-fefdb24f8e10