Don't deal with two test files in src/t/, use only suuid.t
$ git rm suuid.t && git mv perl-suuid.t suuid.t

I was thinking of having a suuid.t where only succeeding tests were run
so I could check regressions in the C version, but this TODO: { } thing
in Test::More still outputs tons of errors, so it's not much point.
Also, because testcmd() and likecmd() each run three tests (stdout,
stderr and retval), it's impossible with the current setup to specify
which of these tests are expected to fail.

84ea1000-38b0-11e6-9006-02010e0a6634