Create and use suuid_xml() + get_tty() returns "not a tty" + more, 61/238 fail
So let get this straight. This is still a wip branch, you can tell by
the Norwegian some places, sorry about that. I'm not bothering with
splitting up commits and stuff, ignoring the "one change per commit"
rule until it's merged back to master. As long as it compiles and
doesn't break more tests, it's ok.

0d96ea4e-3a97-11e6-8b06-d33981fa59d4