Move rc->hostname check into get_hostname(), 48/238 fail
Also move the FAKE_HOST section to a better place and turn it into an
"else if" that always succeeds. When it was at the old place, test 76
and test 77 would fail because it always returned "fake" now that the rc
stuff happens there.

aea1f24a-3cb2-11e6-97d4-cf930a3beb05