Make Entry.tag an array of char * + remove old tag defs, 34/238 fail
Should maybe make it more elegant so it can store an unlimited (as far
as memory goes) number of tags, but nevermind that for now. I've never
used more than three tags at the same time in the 7 years 9 mons 30 days
this thing has been alive, so the current MAX_TAGS limit of 1000 should
be quite enough.

92043272-3cf0-11e6-94c5-f96f68f107a8