Get rid of create_logfile() and set_up_logfile(), move into open_logfile()
Trying to get rid opf a small race condition when the log file is
created. In some situations it can create an extra header when two
processes creates the file at the same time. The creation isn't atomic,
there's a small gap between the existance check and when it's opened and
locked. It's hard to reproduce, and it's maybe a non-problem because it
only happens when the file is created and lots of processes is hammering
on it, that's an unlikely scenario. Nevertheless, I'll look into it
because it's kind of annoying.

26f9e6ee-431b-11e6-86b7-02010e0a6634