free_sess(): Make sure to free all entry->sess[].desc strings
If some of the desc strings were set to NULL because they either was
invalid or missing, it stopped there without deallocating all of them.
Because all sess entries have an UUID, but not necessarily a
description, loop through all the UUIDs and free() all descriptions
whether they're NULL or not.

Seems as this was the final memory leak. All 309 tests succeed when
running under Valgrind 3.12.0 with the following options:

  valgrind -q --leak-check=full --show-leak-kinds=all

2b082866-a70e-11e6-b979-6a50eec7d3a2