Makefile: Replace gcc with cc
Experimenting with FreeBSD, and there's no gcc there. :-/ On Linux
systems where gcc is installed, it should behave like before, as cc is
just a symlink to gcc. It compiles without warnings on FreeBSD, cool.
Well, that's if you ignore the missing HOST_NAME_MAX, but I'll fix that
in the next commit.

86ced892-a86d-11e6-bc86-525400123456