Move all the preprocessor directives to Makefile, use ifdef/ifndef
Make it possible to build with all preprocessor directives by specifying
for example FAKE_HOST= or FAKEHOST=somevalue (normally 1). Have been
using #if, but seems as it's better to use #ifdef/#ifndef. Also test
various combinations for all directives in "make testall" and "make
valgrindall". The sess executable is always built with FAKE_HOST and
PERL_COMPAT for now, but plans are to eliminate it and move the
functionality into suuid.

1a6a6cfe-b194-11e6-97f9-279c2a0468a3