Don't allow \r (CARRIAGE RETURN (CR), U+000D) in comments
It's been allowed since Day 0 when the very first version was created,
but it's probably a bad idea to open up for EOL issues. The only
accepted characters below U+0020 (SPACE) are \t (CHARACTER TABULATION,
U+0009) and \n (LINE FEED (LF), U+000A) at the moment. It might change
back in the future, but disallow it for now.

96339cc2-b831-11e6-a277-db5caa6d21d3