FAKE_HOST: Always use "fake" as hostname, override SUUID_HOSTNAME and rc
That's how it was done before commit cfa399a5b620 ("Move rc->hostname
check into get_hostname(), 48/238 fail", 2016-06-27). That commit
changed the behaviour to only use "fake" as hostname when it was found
with gethostname() to avoid test failures. Add $FAKE_HOST in suuid.t to
check for fake.xml when it's compiled with FAKE_HOST.

2cd1cdee-e774-11e7-ba6f-f74d993421b0