Don't use HOST_NAME_MAX, create MAX_HOSTNAME_LENGTH with fixed length
The "SUUID_HOSTNAME is too long, 100 bytes" test failed on some
platforms, for example in Termux on Android and some of the BSDs.
Because the hostname is used in the log filename, it's enough to limit
the length to fit into a regular filename. Create MAX_HOSTNAME_LENGTH
and set it to 100, it should work everywhere with a decent file system.

ae784af4-453c-11e9-95d1-4f45262dc9b5