get_current_time(): Don't check for later timestamp, only different
Could be a problem if the -n/--count option is used and it's running for
a long time, for example when creating millions of uuids. If the system
clock is adjusted backwards by ntpd(8) or something else, it could abort
if the count variable reaches one million before the clock catches up to
the previous timestamp. Only check that the current timestamp is
different from the previous one, it doesn't have to be later.

709774b8-46c1-11e9-af60-4f45262dc9b5