Delete sqlite3.c and sqlite3.h, not used
There were some experiments on some branches to store the data in a
SQLite database, but it seems as using plain old XML files is the way to
go. It's been working great for ten years and eleven days and is a safe
storage method that works great with version control.

Another big bonus is that all GNUisms in Makefile are gone now, it
should work with BSD make and other common variants. No need for gmake
on *BSD anymore.

ff0a88b2-4819-11e9-a49c-db5caa6d21d3