Change arg of isspace(), isdigit() and tolower() from int to unsigned char
The test

  not ok 276 - Log contents OK after 3-byte UTF-8 sequence

failed on NetBSD 7.0.2 (gcc 4.8.4) because the argument to isspace() was
cast to int in commit 91dfd045ca79 ("Make it build on NetBSD 7.0.2, cast
*char to int", 2016-11-17), the argument was defined as such in the
manpage. But isspace(3) and friends expect the argument to be
represented as an unsigned char, otherwise the behaviour is undefined.
In this case isspace() in trim_str_front() and trim_str_end()
interpreted byte 1 and 3 of the U+0B72 UTF-8 sequence "\xE0\xAD\xB2" as
whitespace, so it ended up in the logfile as \xAD.

ddd6fb2a-5580-11e9-a10c-4f45262dc9b5