--selftest executes opt_selftest() instead of selftest(), 82:512
The plan is to convert all Perl tests in src/t/ into equivalents in C,
executed with --selftest.

For the next commits while this is going on, the output from `make tlok`
is stored in the commit subject line. The format is "a:b" where `a` is
the number of untested lines and `b` is the number of tests at the time
of commit.

39572ffe-dead-11ef-acfb-754a1be59913