Add test category selection to --selftest and --valgrind
Make it possible to specify which test categories to run using an
optional argument to --selftest and --valgrind. The argument can contain
"exec" for executable tests, "func" for function tests, or "all"
(default). Multiple values can be combined.

Also, when using `make`, the `make` variable `TESTS` can be used to
specify the optional --selftest or --valgrind argument directly.

38e0637e-1cd3-11f0-8446-83850402c3ce