Delete mymalloc() and mystrdup(), use the new failed() macro
It's nice and easy to have mymalloc() and mystrdup() to take care of the
error reporting, but it's harder to pinpoint where the error actually
happened. Replace all mymalloc() and mystrdup() with regular malloc()
and strdup(), and use the new failed() macro for error messages.

e33c9896-2751-11f0-a1a5-83850402c3ce