`opt` isn't global anymore, declare it as `static`
Create `const struct Options *` parameter for all functions that need
access to the members of `opt`. This access is read-only, but for local
modifications, a copy of `opt` can be stored in a local variable with
opt_struct(). An example of this is in test_valgrind_option() where the
command that checks if Valgrind is installed is executed without
Valgrind.

Also create set_opt_valgrind() for use in test_valgrind_option() so it's
able to set `opt.valgrind` to `false` if --valgrind is used and Valgrind
isn't installed.

066c9ab0-38c7-11f0-a697-83850402c3ce