selftest.c: Use EXECSTR in tests containing `argv[0]` output
As described in previous commit `51c74e952ba0` ("test_command(): Replace
text in stdout and stderr with `argv[0]` value"), test_command() now
replaces `__EXSTR__` in stdout and stderr with the value of `argv[0]`.
This enables the use of tc() (full stderr and stdout verification)
instead of just substring verification with sc() in those tests.

92b1b3da-56cb-11f0-a17f-83850402c3ce