Makefile: longlines: Use Perl for consistent UTF-8 handling
`make longlines` (check for lines longer than 79 characters in the
source code) gives false positives with grep(1) on OpenBSD 7.7 (grep
version 0.9) because it's not UTF-8 aware. It fails even when using a
UTF-8 locale (e.g., `LC_CTYPE=en_US.UTF-8`).

awk(1) on Debian-based systems doesn't understand UTF-8 either. The
default `mawk` there doesn't handle UTF-8 well, requiring `gawk` to be
installed instead.

Rewording the text to avoid the problem could be an option, but it's
2025 and a matter of principle at this point. UTF-8 has been the
dominant encoding for around 20 years; it's about time the software
adapts.

So, because we don't know if the installed `grep` or `awk` is retarded,
use Perl instead. It's everywhere and handles UTF-8 correctly.

2b21a7dc-56d2-11f0-b012-83850402c3ce