Don't check errno values unless compiled with CHECK_ERRNO
When `make testold` was executed on Termux, OpenBSD, or FreeBSD, 1 test
that redirects the output to /dev/null failed. When a program is
redirected to /dev/null, errno is strangely enough set to ENOTTY
(Inappropriate ioctl for device).

errno isn't set when redirecting stdout to an ordinary file, so there's
something special about /dev/null.

This is a bug in BSD systems, change my mind.

976a1b1a-5af9-11f0-88c1-f7f8c6927076