process_uuid(): Remove `#ifdef` that resets errno on Termux
This ifdef was added in commit `29f836f4b708` ("process_uuid(): Fix
errno in Termux when redirecting output to /dev/null", 2025-04-23) to
fix a problem on Termux. But because the problem seems to exist on all
BSD variants, all use of `check_errno` was disabled by default in the
previous commit. It's not necessary anymore, so delete it.

5a2956a6-5afb-11f0-ac6e-f7f8c6927076