Add and remove `gncov` markers, 872:178
Add gncov markers to unreachable lines with the current code flow, but
the lines can't be removed in case things change in the future or
something fails. gncov markers in unused code (e.g., `#ifdef UNUSED`)
are also removed, but they will come back if the code is ever used.

157e2d58-7462-11f0-a374-83850402c3ce