Add line numbers to the test output and create `OK_*_L()` macros, 870:194
The test output now prefixes each line with the line number the test was
executed from. Also add `OK_*_L()` macros that are the same as the
regular `OK_*()` macros, but with an extra parameter where the line
number can be specified to communicate to helper functions the line
number where the test was actually executed from. Without this
functionality, it would show the line number inside the helper function,
and that's not very useful.

Commit ID: ef07f190-7727-11f0-a058-83850402c3ce