Add tests for get_sess_info() + make `sessind` file-static, 654:309
`sessind` in sessvar.c had to be reset between each call to fill_sess(),
otherwise it stored the new values further out in the array and the
elements before the new data were NULL. The result of that was that
create_sess_xml() returned an empty string. The reset happens in
free_sess(), and that's indeed a thing that belongs there.

Commit ID: e91d0588-772f-11f0-9a34-83850402c3ce