Delete all FIXMEs about slash in filesystem entries
All uses of slash ('/') in filesystem entries were marked in case the
program was ported to OSes that don't use a slash as directory
separator. For example OpenVMS, Stratus VOS and MS Windows/MS-DOS.
Chances are small it will be ported to any of these OSes. But if it is,
the compiler will probably convert the slash according to local
conventions.

Commit ID: 7be24868-7ba1-11f0-997e-83850402c3ce